move remove button
