+++ deployables/extension/src/panel/pages/routes/list/RemoveButton.tsx
